+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### Fixed
+
+- index: Fixed crash when adding an index with new subscriptions in runtime.
+
 ## [5.2.4] - 2022-07-17
 
 ### Fixed