+++ CHANGELOG.md
@@ -7,10 +7,16 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+### Added
+
 - cli: Added `config export --full` flag to resolve templates before printing config.
 - config: Added `advanced.rollback_depth` field, a number of levels to keep in a database for rollback.
 - context: Added `rollback` method to perform database rollback.
 - database: Added an internal `ModelUpdate` model to store the latest database changes.
+
+### Fixed
+
+- prometheus: Fixed updating `dipdup_index_handlers_matched_total` metric.
 
 ### Changed